ajustes

+++ Espades.Domain/Contracts/Services/Base/IService.cs
@@ -9,8 +9,8 @@ namespace Espades.Domain.Contracts.Services.Base
         where TEntity : BaseEntity
     {
         Task<RequestResult> Save(TEntity entity);
-        Task<RequestResult> Delete(long id);
-        Task<RequestResult> Get(long id);
+        Task<RequestResult> Delete(int id);
+        Task<RequestResult> Get(int id);
         Task<RequestResult> GetAll();
         Task<RequestResult> GetByFilter(FilterHelper filter);
     }
